+++ darts/timeseries.py
@@ -77,6 +77,7 @@ VALID_INDEX_TYPES = (pd.DatetimeIndex, pd.RangeIndex)
 STATIC_COV_TAG = "static_covariates"
 DEFAULT_GLOBAL_STATIC_COV_NAME = "global_components"
 HIERARCHY_TAG = "hierarchy"
+METADATA_TAG = "metadata"
 
 
 class TimeSeries:
@@ -320,8 +321,27 @@ class TimeSeries:
                 c for c in self.components if c in bottom_level
             ]
 
-        # store static covariates and hierarchy in attributes (potentially storing None)
-        self._xa = _xarray_with_attrs(self._xa, static_covariates, hierarchy)
+        # prepare metadata
+        metadata = self._xa.attrs.get(METADATA_TAG, None)
+        if not (
+            isinstance(metadata, pd.Series)
+            or (isinstance(metadata, pd.DataFrame) and metadata.shape[0] == 1)
+            or metadata is None
+        ):
+            raise_log(
+                ValueError(
+                    "`metadata` must be either a pandas Series, 1-rowed DataFrame or None"
+                ),
+                logger,
+            )
+        if metadata is not None:
+            metadata = (
+                metadata.iloc[0] if isinstance(metadata, pd.DataFrame) else metadata
+            )
+            metadata = pd.Series(metadata, name=METADATA_TAG)
+
+        # store static covariates, hierarchy and metadata in attributes (potentially storing None)
+        self._xa = _xarray_with_attrs(self._xa, static_covariates, hierarchy, metadata)
 
     """
     Factory Methods
@@ -481,6 +501,7 @@ class TimeSeries:
         fillna_value: Optional[float] = None,
         static_covariates: Optional[Union[pd.Series, pd.DataFrame]] = None,
         hierarchy: Optional[dict] = None,
+        metadata: Optional[Union[pd.Series, pd.DataFrame]] = None,
         **kwargs,
     ) -> Self:
         """
@@ -544,6 +565,8 @@ class TimeSeries:
             The hierarchy can be used to reconcile forecasts (so that the sums of the forecasts at
             different levels are consistent), see `hierarchical reconciliation
             <https://unit8co.github.io/darts/generated_api/darts.dataprocessing.transformers.reconciliation.html>`_.
+        metadata
+            Optionally, a pandas Series or 1-rowed DataFrame defining properties for metadata attributes.
 
         **kwargs
             Optional arguments to be passed to `pandas.read_csv` function
@@ -564,6 +587,7 @@ class TimeSeries:
             fillna_value=fillna_value,
             static_covariates=static_covariates,
             hierarchy=hierarchy,
+            metadata=metadata,
         )
 
     @classmethod
@@ -577,6 +601,7 @@ class TimeSeries:
         fillna_value: Optional[float] = None,
         static_covariates: Optional[Union[pd.Series, pd.DataFrame]] = None,
         hierarchy: Optional[dict] = None,
+        metadata: Optional[Union[pd.Series, pd.DataFrame]] = None,
     ) -> Self:
         """
         Build a deterministic TimeSeries instance built from a selection of columns of a DataFrame.
@@ -643,6 +668,8 @@ class TimeSeries:
             The hierarchy can be used to reconcile forecasts (so that the sums of the forecasts at
             different levels are consistent), see `hierarchical reconciliation
             <https://unit8co.github.io/darts/generated_api/darts.dataprocessing.transformers.reconciliation.html>`_.
+        metadata
+            Optionally, a pandas Series or 1-rowed DataFrame defining properties for metadata attributes.
 
         Returns
         -------
@@ -736,7 +763,11 @@ class TimeSeries:
             series_df.values[:, :, np.newaxis],
             dims=(time_index.name,) + DIMS[-2:],
             coords={time_index.name: time_index, DIMS[1]: series_df.columns},
-            attrs={STATIC_COV_TAG: static_covariates, HIERARCHY_TAG: hierarchy},
+            attrs={
+                STATIC_COV_TAG: static_covariates,
+                HIERARCHY_TAG: hierarchy,
+                METADATA_TAG: metadata,
+            },
         )
 
         return cls.from_xarray(
@@ -754,6 +785,7 @@ class TimeSeries:
         time_col: Optional[str] = None,
         value_cols: Optional[Union[list[str], str]] = None,
         static_cols: Optional[Union[list[str], str]] = None,
+        metadata_cols: Optional[Union[list[str], str]] = None,
         fill_missing_dates: Optional[bool] = False,
         freq: Optional[Union[str, int]] = None,
         fillna_value: Optional[float] = None,
@@ -794,6 +826,9 @@ class TimeSeries:
             appended as static covariates to the resulting TimeSeries groups. Different to `group_cols`, the
             DataFrame is not grouped by these columns. Note that for every group, there must be exactly one
             unique value.
+        metadata_cols
+            A string or list of strings representing metadata columns from the DataFrame that should be
+            appended as metadata to the resulting TimeSeries groups.
         fill_missing_dates
             Optionally, a boolean value indicating whether to fill missing dates (or indices in case of integer index)
             with NaN values. This requires either a provided `freq` or the possibility to infer the frequency from the
@@ -865,7 +900,25 @@ class TimeSeries:
             value_cols = df.columns.drop(static_cov_cols + extract_time_col).tolist()
         extract_value_cols = [value_cols] if isinstance(value_cols, str) else value_cols
 
-        df = df[static_cov_cols + extract_value_cols + extract_time_col]
+        if metadata_cols is not None:
+            metadata_cols = (
+                [metadata_cols]
+                if not isinstance(metadata_cols, list)
+                else metadata_cols
+            )
+        else:
+            metadata_cols = []
+        # columns that are used for metadata but not for grouping or static covariates
+        extract_metadata_cols = [
+            col for col in metadata_cols if col not in static_cov_cols
+        ]
+
+        df = df[
+            static_cov_cols
+            + extract_value_cols
+            + extract_time_col
+            + extract_metadata_cols
+        ]
 
         if time_col:
             if np.issubdtype(df[time_col].dtype, object) or np.issubdtype(
@@ -889,6 +942,7 @@ class TimeSeries:
 
         groups = df.groupby(group_cols[0] if len(group_cols) == 1 else group_cols)
 
+        # build progress bar for iterator
         iterator = _build_tqdm_iterator(
             groups,
             verbose=verbose,
@@ -937,6 +991,29 @@ class TimeSeries:
                 # add the static covariates to the group values
                 static_cov_vals += tuple(group[static_cols].values[0])
 
+            metadata_vals = tuple()
+            # check that for each group there is only one unique value per column in `metadata_cols`
+            if metadata_cols:
+                metadata_cols_valid = [
+                    len(group[col].unique()) == 1 for col in metadata_cols
+                ]
+                if not all(metadata_cols_valid):
+                    invalid_cols = [
+                        metadata_col
+                        for metadata_col, is_valid in zip(
+                            metadata_cols, metadata_cols_valid
+                        )
+                        if not is_valid
+                    ]
+                    raise_if(
+                        True,
+                        f"Encountered more than one unique value in group {group} for given metadata columns: "
+                        f"{invalid_cols}.",
+                        logger,
+                    )
+                # add the metadata to the group values
+                metadata_vals += tuple(group[metadata_cols].values[0])
+
             return cls.from_dataframe(
                 df=split,
                 fill_missing_dates=fill_missing_dates,
@@ -945,6 +1022,11 @@ class TimeSeries:
                 static_covariates=(
                     pd.DataFrame([static_cov_vals], columns=extract_static_cov_cols)
                     if extract_static_cov_cols
+                    else None
+                ),
+                metadata=(
+                    pd.DataFrame([metadata_vals], columns=metadata_cols)
+                    if metadata_cols
                     else None
                 ),
             )
@@ -965,6 +1047,7 @@ class TimeSeries:
         freq: Optional[Union[str, int]] = None,
         fillna_value: Optional[float] = None,
         static_covariates: Optional[Union[pd.Series, pd.DataFrame]] = None,
+        metadata: Optional[Union[pd.Series, pd.DataFrame]] = None,
     ) -> Self:
         """
         Build a univariate deterministic series from a pandas Series.
@@ -995,6 +1078,8 @@ class TimeSeries:
             Optionally, a set of static covariates to be added to the TimeSeries. Either a pandas Series or a
             single-row pandas DataFrame. If a Series, the index represents the static variables. If a DataFrame, the
             columns represent the static variables and the single row represents the univariate TimeSeries component.
+        metadata
+            Optionally, a pandas Series or 1-rowed DataFrame defining properties for metadata attributes.
 
         Returns
         -------
@@ -1010,6 +1095,7 @@ class TimeSeries:
             freq=freq,
             fillna_value=fillna_value,
             static_covariates=static_covariates,
+            metadata=metadata,
         )
 
     @classmethod
@@ -1023,6 +1109,7 @@ class TimeSeries:
         fillna_value: Optional[float] = None,
         static_covariates: Optional[Union[pd.Series, pd.DataFrame]] = None,
         hierarchy: Optional[dict] = None,
+        metadata: Optional[Union[pd.Series, pd.DataFrame]] = None,
     ) -> Self:
         """
         Build a series from a time index and value array.
@@ -1085,6 +1172,8 @@ class TimeSeries:
             The hierarchy can be used to reconcile forecasts (so that the sums of the forecasts at
             different levels are consistent), see `hierarchical reconciliation
             <https://unit8co.github.io/darts/generated_api/darts.dataprocessing.transformers.reconciliation.html>`_.
+        metadata
+            Optionally, a pandas Series or 1-rowed DataFrame defining properties for metadata attributes.
 
         Returns
         -------
@@ -1123,7 +1212,11 @@ class TimeSeries:
             values,
             dims=(times_name,) + DIMS[-2:],
             coords=coords,
-            attrs={STATIC_COV_TAG: static_covariates, HIERARCHY_TAG: hierarchy},
+            attrs={
+                STATIC_COV_TAG: static_covariates,
+                HIERARCHY_TAG: hierarchy,
+                METADATA_TAG: metadata,
+            },
         )
         return cls.from_xarray(
             xa=xa,
@@ -1140,6 +1233,7 @@ class TimeSeries:
         fillna_value: Optional[float] = None,
         static_covariates: Optional[Union[pd.Series, pd.DataFrame]] = None,
         hierarchy: Optional[dict] = None,
+        metadata: Optional[Union[pd.Series, pd.DataFrame]] = None,
     ) -> Self:
         """
         Build an integer-indexed series from an array of values.
@@ -1187,6 +1281,8 @@ class TimeSeries:
             The hierarchy can be used to reconcile forecasts (so that the sums of the forecasts at
             different levels are consistent), see `hierarchical reconciliation
             <https://unit8co.github.io/darts/generated_api/darts.dataprocessing.transformers.reconciliation.html>`_.
+        metadata
+            Optionally, a pandas Series or 1-rowed DataFrame defining properties for metadata attributes.
 
         Returns
         -------
@@ -1207,6 +1303,7 @@ class TimeSeries:
             fillna_value=fillna_value,
             static_covariates=static_covariates,
             hierarchy=hierarchy,
+            metadata=metadata,
         )
 
     @classmethod
@@ -1215,6 +1312,7 @@ class TimeSeries:
         json_str: str,
         static_covariates: Optional[Union[pd.Series, pd.DataFrame]] = None,
         hierarchy: Optional[dict] = None,
+        metadata: Optional[Union[pd.Series, pd.DataFrame]] = None,
     ) -> Self:
         """
         Build a series from the JSON String representation of a ``TimeSeries``
@@ -1258,6 +1356,8 @@ class TimeSeries:
             The hierarchy can be used to reconcile forecasts (so that the sums of the forecasts at
             different levels are consistent), see `hierarchical reconciliation
             <https://unit8co.github.io/darts/generated_api/darts.dataprocessing.transformers.reconciliation.html>`_.
+        metadata
+            Optionally, a pandas Series or 1-rowed DataFrame defining properties for metadata attributes.
 
         Returns
         -------
@@ -1266,7 +1366,10 @@ class TimeSeries:
         """
         df = pd.read_json(StringIO(json_str), orient="split")
         return cls.from_dataframe(
-            df, static_covariates=static_covariates, hierarchy=hierarchy
+            df,
+            static_covariates=static_covariates,
+            hierarchy=hierarchy,
+            metadata=metadata,
         )
 
     @classmethod
@@ -1317,6 +1420,15 @@ class TimeSeries:
         and values are the set of parent(s) of these components in the hierarchy.
         """
         return self._xa.attrs.get(HIERARCHY_TAG, None)
+
+    @property
+    def metadata(self) -> Optional[pd.Series]:
+        """
+        The metadata of this TimeSeries, if any.
+        If set, the metadata is encoded as a pandas Series which defines properties
+        for identifying and describing the nature of the TimeSeries.
+        """
+        return self._xa.attrs.get(METADATA_TAG, None)
 
     @property
     def has_hierarchy(self) -> bool:
@@ -2028,6 +2140,7 @@ class TimeSeries:
         ignore_time_axis: Optional[bool] = False,
         ignore_static_covariates: bool = False,
         drop_hierarchy: bool = True,
+        drop_metadata: bool = True,
     ) -> Self:
         """
         Concatenate another timeseries to the current one along given axis.
@@ -2050,7 +2163,9 @@ class TimeSeries:
             (by merging the hierarchy dictionaries), which may cause issues if the component
             names of the resulting series and that of the merged hierarchy do not match.
             When `axis=0` or `axis=2`, the hierarchy of the first series is always kept.
-
+        drop_metadata : bool
+            Whether to drop the metadata information of the concatenated series. True by default.
+            When False, the concatenated series will inherit the metadata from the current timeseries.
 
         Returns
         -------
@@ -2072,6 +2187,7 @@ class TimeSeries:
             ignore_time_axis=ignore_time_axis,
             ignore_static_covariates=ignore_static_covariates,
             drop_hierarchy=drop_hierarchy,
+            drop_metadata=drop_metadata,
         )
 
     """
@@ -2612,6 +2728,7 @@ class TimeSeries:
             columns=self.components,
             static_covariates=self.static_covariates,
             hierarchy=self.hierarchy,
+            metadata=self.metadata,
         )
 
     def longest_contiguous_slice(
@@ -2927,6 +3044,7 @@ class TimeSeries:
                 times=idx,
                 fill_missing_dates=False,
                 static_covariates=self.static_covariates,
+                metadata=self.metadata,
             )
         )
 
@@ -2999,6 +3117,7 @@ class TimeSeries:
                 static_covariates=self.static_covariates,
                 columns=self.columns,
                 hierarchy=self.hierarchy,
+                metadata=self.metadata,
             )
         )
 
@@ -3058,6 +3177,7 @@ class TimeSeries:
             fillna_value=fillna_value,
             static_covariates=self.static_covariates,
             hierarchy=self.hierarchy,
+            metadata=self.metadata,
         )
 
     def with_values(self, values: np.ndarray) -> Self:
@@ -3148,6 +3268,7 @@ class TimeSeries:
                 attrs={
                     STATIC_COV_TAG: covariates,
                     HIERARCHY_TAG: self.hierarchy,
+                    METADATA_TAG: self.metadata,
                 },
             )
         )
@@ -3190,6 +3311,32 @@ class TimeSeries:
                 attrs={
                     STATIC_COV_TAG: self.static_covariates,
                     HIERARCHY_TAG: hierarchy,
+                    METADATA_TAG: self.metadata,
+                },
+            )
+        )
+
+    def with_metadata(self, metadata: Optional[Union[pd.Series, pd.DataFrame]]):
+        """
+        Adds metadata to the TimeSeries.
+
+        Parameters
+        ----------
+        metadata
+            A pandas Series or 1-rowed DataFrame defining properties for identifying and describing
+            the underlying data of the TimeSeries. The types of the values can be different. For example:
+            ``metadata = pd.Series({'source_file': 'woolyrnq.csv', 'has_wool': True, 'start_year': 1965})``
+        """
+
+        return self.__class__(
+            xr.DataArray(
+                self._xa.values,
+                dims=self._xa.dims,
+                coords=self._xa.coords,
+                attrs={
+                    STATIC_COV_TAG: self.static_covariates,
+                    HIERARCHY_TAG: self.hierarchy,
+                    METADATA_TAG: metadata,
                 },
             )
         )
@@ -4988,7 +5135,7 @@ class TimeSeries:
     def __add__(self, other):
         if isinstance(other, (int, float, np.integer)):
             xa_ = _xarray_with_attrs(
-                self._xa + other, self.static_covariates, self.hierarchy
+                self._xa + other, self.static_covariates, self.hierarchy, self.metadata
             )
             return self.__class__(xa_)
         elif isinstance(other, (TimeSeries, xr.DataArray, np.ndarray)):
@@ -5007,7 +5154,7 @@ class TimeSeries:
     def __sub__(self, other):
         if isinstance(other, (int, float, np.integer)):
             xa_ = _xarray_with_attrs(
-                self._xa - other, self.static_covariates, self.hierarchy
+                self._xa - other, self.static_covariates, self.hierarchy, self.metadata
             )
             return self.__class__(xa_)
         elif isinstance(other, (TimeSeries, xr.DataArray, np.ndarray)):
@@ -5026,7 +5173,7 @@ class TimeSeries:
     def __mul__(self, other):
         if isinstance(other, (int, float, np.integer)):
             xa_ = _xarray_with_attrs(
-                self._xa * other, self.static_covariates, self.hierarchy
+                self._xa * other, self.static_covariates, self.hierarchy, self.metadata
             )
             return self.__class__(xa_)
         elif isinstance(other, (TimeSeries, xr.DataArray, np.ndarray)):
@@ -5046,7 +5193,10 @@ class TimeSeries:
         if isinstance(n, (int, float, np.integer)):
             raise_if(n < 0, "Attempted to raise a series to a negative power.", logger)
             xa_ = _xarray_with_attrs(
-                self._xa ** float(n), self.static_covariates, self.hierarchy
+                self._xa ** float(n),
+                self.static_covariates,
+                self.hierarchy,
+                self.metadata,
             )
             return self.__class__(xa_)
         if isinstance(n, (TimeSeries, xr.DataArray, np.ndarray)):
@@ -5064,7 +5214,7 @@ class TimeSeries:
             if other == 0:
                 raise_log(ZeroDivisionError("Cannot divide by 0."), logger)
             xa_ = _xarray_with_attrs(
-                self._xa / other, self.static_covariates, self.hierarchy
+                self._xa / other, self.static_covariates, self.hierarchy, self.metadata
             )
             return self.__class__(xa_)
         elif isinstance(other, (TimeSeries, xr.DataArray, np.ndarray)):
@@ -5107,13 +5257,14 @@ class TimeSeries:
     def __lt__(self, other) -> xr.DataArray:
         if isinstance(other, (int, float, np.integer, np.ndarray, xr.DataArray)):
             return _xarray_with_attrs(
-                self._xa < other, self.static_covariates, self.hierarchy
+                self._xa < other, self.static_covariates, self.hierarchy, self.metadata
             )
         elif isinstance(other, TimeSeries):
             return _xarray_with_attrs(
                 self._xa < other.data_array(copy=False),
                 self.static_covariates,
                 self.hierarchy,
+                self.metadata,
             )
         else:
             raise_log(
@@ -5126,13 +5277,14 @@ class TimeSeries:
     def __gt__(self, other) -> xr.DataArray:
         if isinstance(other, (int, float, np.integer, np.ndarray, xr.DataArray)):
             return _xarray_with_attrs(
-                self._xa > other, self.static_covariates, self.hierarchy
+                self._xa > other, self.static_covariates, self.hierarchy, self.metadata
             )
         elif isinstance(other, TimeSeries):
             return _xarray_with_attrs(
                 self._xa > other.data_array(copy=False),
                 self.static_covariates,
                 self.hierarchy,
+                self.metadata,
             )
         else:
             raise_log(
@@ -5145,13 +5297,14 @@ class TimeSeries:
     def __le__(self, other) -> xr.DataArray:
         if isinstance(other, (int, float, np.integer, np.ndarray, xr.DataArray)):
             return _xarray_with_attrs(
-                self._xa <= other, self.static_covariates, self.hierarchy
+                self._xa <= other, self.static_covariates, self.hierarchy, self.metadata
             )
         elif isinstance(other, TimeSeries):
             return _xarray_with_attrs(
                 self._xa <= other.data_array(copy=False),
                 self.static_covariates,
                 self.hierarchy,
+                self.metadata,
             )
         else:
             raise_log(
@@ -5164,13 +5317,14 @@ class TimeSeries:
     def __ge__(self, other) -> xr.DataArray:
         if isinstance(other, (int, float, np.integer, np.ndarray, xr.DataArray)):
             return _xarray_with_attrs(
-                self._xa >= other, self.static_covariates, self.hierarchy
+                self._xa >= other, self.static_covariates, self.hierarchy, self.metadata
             )
         elif isinstance(other, TimeSeries):
             return _xarray_with_attrs(
                 self._xa >= other.data_array(copy=False),
                 self.static_covariates,
                 self.hierarchy,
+                self.metadata,
             )
         else:
             raise_log(
@@ -5331,6 +5485,7 @@ class TimeSeries:
                         else xa_.attrs[STATIC_COV_TAG]
                     ),
                     None,
+                    xa_.attrs[METADATA_TAG],
                 )
                 return self.__class__(xa_)
             elif isinstance(key.start, (int, np.int64)) or isinstance(
@@ -5366,6 +5521,7 @@ class TimeSeries:
                     else xa_.attrs[STATIC_COV_TAG]
                 ),
                 None,
+                xa_.attrs[METADATA_TAG],
             )
             return self.__class__(xa_)
         elif isinstance(key, (int, np.int64)):
@@ -5407,6 +5563,7 @@ class TimeSeries:
                         else xa_.attrs[STATIC_COV_TAG]
                     ),
                     None,
+                    xa_.attrs[METADATA_TAG],
                 )
                 return self.__class__(xa_)
             elif all(isinstance(i, (int, np.int64)) for i in key):
@@ -5448,13 +5605,14 @@ class TimeSeries:
         raise_log(IndexError("The type of your index was not matched."), logger)
 
 
-def _xarray_with_attrs(xa_, static_covariates, hierarchy):
+def _xarray_with_attrs(xa_, static_covariates, hierarchy, metadata):
     """Return an DataArray instance with static covariates and hierarchy stored in the array's attributes.
     Warning: This is an inplace operation (mutable) and should only be called from within TimeSeries construction
-    or to restore static covariates and hierarchy after operations in which they did not get transferred.
+    or to restore static covariates, hierarchy and metadata after operations in which they did not get transferred.
     """
     xa_.attrs[STATIC_COV_TAG] = static_covariates
     xa_.attrs[HIERARCHY_TAG] = hierarchy
+    xa_.attrs[METADATA_TAG] = metadata
     return xa_
 
 
@@ -5533,6 +5691,7 @@ def concatenate(
     ignore_time_axis: bool = False,
     ignore_static_covariates: bool = False,
     drop_hierarchy: bool = True,
+    drop_metadata: bool = True,
 ):
     """Concatenates multiple ``TimeSeries`` along a given axis.
 
@@ -5560,6 +5719,9 @@ def concatenate(
         "concatenated" as well (by merging the hierarchy dictionaries), which may cause issues if the component
         names of the resulting series and that of the merged hierarchy do not match.
         When `axis=0` or `axis=2`, the hierarchy of the first series is always kept.
+    drop_metadata : bool
+        Whether to drop the metadata information of the concatenated series. True by default.
+        When False, the concatenated series will inherit the metadata from the first TimeSeries element in `series`.
 
     Returns
     -------
@@ -5589,6 +5751,8 @@ def concatenate(
 
     component_axis_equal = len({ts.width for ts in series}) == 1
     sample_axis_equal = len({ts.n_samples for ts in series}) == 1
+
+    metadata = None if drop_metadata else series[0].metadata
 
     if axis == 0:
         # time
@@ -5624,7 +5788,7 @@ def concatenate(
 
             da_concat = da_concat.assign_coords({time_dim_name: tindex})
             da_concat = _xarray_with_attrs(
-                da_concat, series[0].static_covariates, series[0].hierarchy
+                da_concat, series[0].static_covariates, series[0].hierarchy, metadata
             )
 
     else:
@@ -5680,7 +5844,11 @@ def concatenate(
             concat_vals,
             dims=(time_dim_name,) + DIMS[-2:],
             coords={time_dim_name: series[0].time_index, DIMS[1]: component_index},
-            attrs={STATIC_COV_TAG: static_covariates, HIERARCHY_TAG: hierarchy},
+            attrs={
+                STATIC_COV_TAG: static_covariates,
+                HIERARCHY_TAG: hierarchy,
+                METADATA_TAG: metadata,
+            },
         )
 
     return TimeSeries.from_xarray(da_concat, fill_missing_dates=False)